Release version 3.3.6

+++ src/zimpl/strstore2.c
@@ -92,7 +92,7 @@ const char* str_new(const char* s)
    if (len > MAX_STR_STORE_SIZE)
    {
       fprintf(stderr, "*** Error 803: String too long %zu > %zu\n",
-         len + 1, MAX_STR_STORE_SIZE); 
+         len + 1, (size_t)MAX_STR_STORE_SIZE); 
 
       zpl_exit(EXIT_FAILURE);
    }
